Add ability to speak character when cursor routing (#16947)
Closes #8072

Summary of the issue:
When cursor routing, it can sometimes be handy if speech follows along and tells you what character you moved to, especially when the character in braille is unknown to you.

Description of user facing changes
Added a new braille option "Speak character when routing cursor in text", disabled by default. When enabled, NVDA will speak the character you routed to after a cursor routing action.
Description of development approach
Added a new option to the config spec, as well as a function to braille to speak the character at the given position. Also added an item to the settings panel as well as an unbound global command.