Set up assign milestone to use base environment variables, not branch (#16974)
Fixup #16779

Summary of the issue:
The GitHub actions to assign milestones to closed issues and PRs is failing for community PRs.
This is because the action is looking in the forks repository for the milestone ID environment variable.

Description of user facing changes
Fixes assigning milestones to community contributions

Description of development approach
using the pull_request_target event rather than pull_request, uses the target (i.e. this) repositories environment variables.