Support multi-line braille displays in HidBrailleDriver (#17001)
Closes #16993

Summary of the issue:
Multi-line braille displays using the HID standard driver only support the first line of cells.

Description of user facing changes
Users with multi-line braille displays that connect using the HidBrailleDriver will now be able to use all cells of their display.

Description of development approach
HidBrailleDriver._cellValueCaps is now a list of all braille cell controls in the device. numRows is set to the length of that list and numCols is set to the ReportCount of each element of the list. In display, the cells are split across the elements of _cellValueCaps.