Removed \userData` from default `configFolder` (#17013)
Summary of the issue:
The `tests/manual/createUpdatableAddons.ps1` script exists to expedite testing the add-on store. It provides defaults for addon name, NVDA config directory and addon major and minor versions. However, the default value for the config directory of an installed copy of NVDA is set incorrectly.

Description of user facing changes
None. This will only impact developers testing the add-on store.

Description of development approach
Updated `configFolder` to remove the `userData` subdirectory, which is not used in installed copies of NVDA.

Testing strategy:
Tested the script after the change to make sure it correctly updated the add-on store metadata and add-on manifest for the Clock add-on.