Update installer to delete `nvda_eoaProxy.exe` (#17023)
Summary of the issue:
NVDA dropped Windows 7 support with NVDA 2024.1. The installer should delete executables that were used by older versions of NVDA but which are now obsolete. However, this was not done for nvda_eoaProxy.exe, which existed to support Windows Vista/7 ease of access.

Description of user facing changes
None.

Description of development approach
Added nvda_eoaProxy.exe to the list of executables to be deleted by the installer.