Do not complete add-on installation or removal from the launcher (#17024)
Fixes #16837

Summary of the issue:
As reported by @CyrilleB79, if one were to update an add-on from the add-on store, then accept an NVDA update before restarting to complete the add-on update, the update will completely clobber the add-on. Upon further investigation, this is because the launcher erroneously attempts to complete add-on installation/removal, but an error is thrown (because it shouldn't be touching the file system) part way through, so it cannot complete successfully.

Description of user facing changes
If add-ons are pending installation/removal when an update is started, the installation will not affect this, and the installation/removal will proceed when the updated copy of NVDA starts.

Description of development approach
Added checks for NVDAState.shouldWriteToDisk() to sections of addonHandler._getAvailableAddonsFromPath that alter the filesystem.