Support compiling with VS 2022 V17.11 (#17028)
Fixes #17026

Summary of the issue:
Trying to compile NVDA with Visual Studio 2022 V17.11 produces the following error:

build\x86\local\beeps.cpp(19): error C2039: 'min': is not a member of 'std'
std::min and std::max are defined in <algorithm> which we never include in beeps.cpp. I assume that <cmath> used to include <algorithm> and now it does not.

Description of user facing changes
None.

Description of development approach
Include <algorithm> in beeps.cpp.