Update auto milestone action to only apply to completed issues and merged PRs (#17040)
The current auto milestone script currently applies milestones to all closed issues and pull requests, even when the issues are invalid or pull requests not merged.

Description of user facing changes
None. Purely a DX improvement.

Description of developer facing changes
If an issue is closed as completed, the default milestone will be applied if the issue does not yet have a milestone.
If an issue is closed as not planned, no milestone will be auto-applied.
If a pull request is merged and closed, the default milestone will be applied if the PR does not yet have a milestone.
If a pull request is closed without merging, no milestone will be auto-applied.
Description of development approach
Added a step to the auto milestone script that checks the close reason of the issue, or merge status of the PR. If the issue is closed as complete or PR closed and merged, the isDone output is set to 'true', otherwise it is set to 'false'. Updated the "Assign default milestone" step's if condition to require isDone to be 'true'.