Reformat build guide and add note about building serially when debugging the build process (#17029)
The build guide currently suggests that developers use the --all-cores parameter to SCons to speed up their builds, noting that this will scramble the output. However, it does not mention that in some cases this may break building altogether.

Description of user facing changes
None. Developer change only.

Description of development approach
Updated projectDocs/dev/buildingNVDA.md:

Fixed a bunch of linting issues, including having blanks around headings, using 2 spaces for list indents, and not having duplicate headings.
Added a heading for building NVDA without archiving it.
Noted that the -j N option can be used instead of the --all-cores option.
Noted that using a parallel build may break the build system.
Noted that when experiencing problems with the build system, make sure to run it serially to make finding the source of the issue easier.