CursorManager._caretMovementScriptHelper: return early if another script is waiting. (#17049)
Fixes #16812

Summary of the issue:
If holding down an arrow key in NvDA browse mode, such as in Microsoft Word, NVDA can become unresponsive if holding down the key for a long period of time. This is because moving in browse mode can be very costly, and scripts become backed up.

Description of user facing changes
NVDA no longer becomes unresponsive if holding down an arrow key for a long time while in NVDA browse mode, in particular in Microsoft word.

Description of development approach
CursorManager._caretMovementScriptHelper: return early if another script is waiting. There is no point doing something costly when another script is going to happen anyway.