Work around broken Data validation dropdown lists in Microsoft Excel (#17042)
Fixes #15138

Summary of the issue:
If a cell in Microsoft Excel has a data validation type of list, NvDA does not report anything when it is opened and or when selecting a value.
Note that this is specifically talking about the list shown when the cell has data validation set. The dropdown list shown when auto completing from values in higher cells does report okay.

Technical info
The list is a SysListView32 control.
It has a native UIA implementation.
NvDA generally does not trust UIA implementations for SysListView32 as they have been somewhat limited.
Therefore NVDA falls back to MSAA for the list.
Although UIA core is proxying the UIA events to MSAA events, the UIA implementation seems to be incomplete as AccessibleObjectFromEvent is always returning E_FAIL for any focus win event fired by the control.
Even if NvDA is told to specifically use UIA for this control, there is a loop in the control's parent chain, causing a never ending focus ancestry.
Description of user facing changes
It is now possible tin teract with data validation dropdown lists in Excel. this includes reporting the focused / selected item as you move up and down the list, and reporting focus back on the sheet when the list is closed with enter, space or escape.

Description of development approach
Excel appModule's isGoodUIAWindow treats this SysListView32 window as native UIA.
The parent property on this NvDAObject has been overridden to return the desktop, skipping the loop between SysListView32 and __XLACOOUTER windows.
AScript on this NvDAObject bound to enter, space and escape, corrects NvDA's focus back to the sheet.