Report a user friendly message when the documentation is missing (#17052)
When NVDA is run from source, if the documentation is missing, e.g. because it has not been built, there is a user friendly message indicating that the documentation is missing when using context help. Though, when trying to open the documentation from NVDA's menu, there is no user friendly message; there is only an error in the log.

Description of user facing changes
If the documentation is missing, a message pops up to report it to the user. In this case, we do not use ui.message as for context help, because the message is cut by the new focus being reported.

Description of development approach
Factored the function to report that the doc is missing.

Testing strategy:
Manual test, with and without built doc:
* Try to use context help
* NVDA menu -> Help: try to open User Guide, Key commands and Changes documentations.