Chromium: disable caret at end of line detection as Chromium does not yet implement what we need, and it is costly for no gain. (#17067)
More broadly fixes #17039

Summary of the issue:
In PR #16745, code was added to detect the caret at end of lines in Mozilla-compatible edit areas, so as to not report the next line when on the final insertion point of the line before.
Although this works great in Firefox and other Gecko-based apps, it does not work at all in Chromium due to Chromium not correctly
implementing IAccessibleText::textAtOffset with IA2_OFFSET_CARET.
More importantly, this is a very costly check in VS code, which has a noticeable performance hit when arrowing up and down large files (issue #17039), for no gain.
PR #17051 successfully fixed the performance hit in VS code by disabling caret at end of line detection just for VS Code. But it was pointed out that there are several VS Code variants including VS Code.dev which require this also.

Description of user facing changes
NVDA is no longer as sluggish when arrowing up and down large fles in VS code and otherChromium-based applications.

Description of development approach
Remoe the VSCodeditor and VSCodeEditorTextInfo classes from the VS ode appModule introduced in pr VS Code is no longer as sluggish when arrowing up and down through large files #17051.
Add Editor and EditorTextInfo classes to NVDAObjects.IAccessible.chromium which are used in Chromium where ever NVDAObject.IAccessible.ia2Web.Editor is used. These classes disable caret at end of line detection for any chromium edit area.