Support IAccessible2 "mark" text attribute. (#16910)
Text fragments is a new web concept that allows you to link directly to a specified piece of arbitrary text on a page without needing a predefined id. For example:
https://www.jantrid.net/#:~:text=browsers%20were%20much%20simpler
As well as scrolling to the element containing the text, the browser will highlight the text. Because this could span a partial object, this can't be exposed using IA2_ROLE_MARK. Instead, similar to spelling errors, this is exposed using an IAccessible2 text attribute, specifically the new attribute mark:true.

This will also be used eventually to support CSS custom highlights.

Description of user facing changes
In Mozilla Firefox, NVDA will report the highlighted text when a URL containing a text fragment is visited.

Description of development approach
NVDA already supports the "marked" attribute on FormatFields. To be consistent with IA2_ROLE_MARK, it was decided to name the IA2 attribute "mark" instead of "marked". This PR simply maps the attribute appropriately in normalizeIA2TextFormatField.