Fix translatable string issue in Albatross driver (#17116)
Summary of the issue:
In Albatross driver, a string consisting in the concatenation of normal strings (`"..."`) and an f-string) is passed to the `_` translation function. As a consequence:
* Only a part of the concatenated string appears in the `.pot` file instead of the full concatenation
* In any case, f-strings cannot be used for translatable messages since the formatting should be applied only after the gettext function has been called

Thus, this message was not translated.

Description of user facing changes
The message of the Albatross driver is now translated.

Description of development approach
Do not use f-string. Instead used `.format` function, applied to the translated string. Slightly reworded the returned message for clarity.

Testing strategy:
- Check in the `.pot` file that we have the full message

Known issues with pull request:
None