Add ability to recognize links to the same page (#16994)
Fixes #141

Summary of the issue:
NVDA cannot recognize when link destination points to the same document, and reporting this information is desired by many users.

Description of user facing changes
NVDA can report if a link destination points to the same document.

Description of development approach
A new INTERNAL_LINK state has been added to controlTypes.State.
NVDA object has a new linkType property, used to report internal (same page) links.
If the object is an internal link, the INTERNAL_LINKstate is added in _get_states, in IAccessible/IA2web and in UIA/Chromium.
In the gecko_ia2 virtual buffer backend, accValue is exposed for links.
In virtualBuffers.gecko_ia2.Gecko_ia2_TextInfo._normalizeControlField, the accValue is used to add the linkType state.
A configuration option has been added to enable or disable this feature. This can be changed through gui or with an unassigned gesture.
BrowseModeInterceptor object has a new documentUrl property and a new getLinkTypeInDocument method. This method uses a helper function of a new urlUtils submodule to check the provided URL.