Fix popup menu item reading on Thunderbird (#4708)
Summary of the issue:
When the user finds text on Thunderbird by pressing CONTROL+K, searching the text and activating one of the links, NVDA speaks only "section".

Description of user facing changes
NVDA will speak the popup menu item name correctly.

Description of developer facing changes
None.

Description of development approach
Added a check if the object 's name is empty, and is of the "div" tag and has a class of "popup-menuitem", NVDA replaces its name with the name of its child.

Testing strategy:
Performed a manual test with Thunderbird.