Update codingStandards to remove instruction to double indent function parameters, and to note instructions for nameing scripts. (#17126)
In-response-to #12355 (comment)

Summary of the issue:
Pre-commit-ci, on behalf of Ruff, does not approve of function parameter lists being double indented.
I.e. NVDA's coding standards said:

For the parameter list of function definitions, double indent, this differentiates the parameters and the body of the function.

Updated coding standards.
Also noticed that there was no instruction on how to format script names, but there was for events and various other entities. Added one to the relevant section.
Changed a few "eg"s to "e.g."s, and made other slight grammar alterations.