Fix braille breakage when activating say all with a config profile enabled (#17164)
Fixed #17163

Summary of the issue:
When you have a config profile attached to say all and activate a config profile, braille is broken until you move out and into the document control.

Description of user facing changes
Braille follows along properly.

Description of development approach
The problem was that in the braille handler, self._tether was set to auto when it was configured to auto. However self._tether should only contain the internal tether state, focus or review. Therefore if it is configured as auto, do not override the internal tether state of the braille handler.