Fix config profiles quickly get polluted with untouched settings (#17158)
Fixes #17157
Fixup of #14133
Partially reverts #15074

Summary of the issue:
When saving configuration profiles, unchanged values are yet saved to a profile.

Description of user facing changes
Configuration profiles should become less polluted after this change.

Description of development approach
Restored the behavior of #14133, mostly reverting #15074. There was namely a False assumption in #15074.
@seanbudd wrote:

This caused various issues including config values not being correctly converted to numbers from strings when validating.

In fact, the validation was not the problem here. Instead, when __getitem__ was called with checkValidity set to False, the unvalidated value was yet saved in the cache. Therefore subsequent calls of __getitem__ would always return the unvalidated cached value that would always be of the string type.